split footer in sub components
